src/main/java/com/fastcash/moneytransfer/util/RSAKeyPairGenerator.java
package com.fastcash.moneytransfer.util;

import java.security.KeyPair;
import java.security.KeyPairGenerator;
import java.security.NoSuchAlgorithmException;

import org.springframework.stereotype.Component;

@Component
public class RSAKeyPairGenerator implements LocalKeyPairGenerator {
	
	@Override
	public KeyPair generateKeyPair() throws NoSuchAlgorithmException {
        KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("RSA");
        keyPairGenerator.initialize(2048);
        return keyPairGenerator.generateKeyPair();
    }
	
}